pyproject.toml
[tool.poetry]
name = "django-looking-for-group"
version = "1.11.0"
description = "An RPG group finding and campaign management application."
authors = ["Daniel Andrlik <daniel@andrlik.org>"]

[tool.poetry.dependencies]
python = "~3.6 || ^3.7 || ^3.8"
pytz = "=2019.1"
text-unidecode = "=1.2"
pillow = "=6.0.0"
rcssmin = "~1.0.6"
rjsmin = "~1.1.0"
argon2-cffi = "-19.1.0"
whitenoise = "~4.1.2"
redis = "<3.0"
django-environ = "~0.4.5"
django-model-utils = "~3.1.2"
django-allauth = "~0.39.1"
django-foundation-formtags = "~0.1.0"
django-compressor = "=2.3"
django-libsass = "~0.7"
django-redis = "~4.10.0"
djangorestframework = "~3.10.3"
coreapi = "~2.3.3"
waitress = "~1.3.0"
django-anymail = "~6.0.1"
django-storages = "~1.7.1"
django-avatar = "~4.1.0"
pwned-passwords-django = "~1.3.2"
django-cookie-law = "~2.0.1"
python-stdnum = "=1.11"
django-isbn-field = "~0.5.2"
django-taggit = "~1.1.0"
django-taggit-serializer = "~0.1.7"
django-braces = "~1.13.0"
rules = "^2.1"
django-filter = "~2.1.0"
markdown = "~3.1.1"
bleach = "~3.1.0"
bleach-whitelist = "~0.0.10"
django-star-ratings = "~0.7.0"
django-q = "~1.0.1"
django-extensions = "~2.1.9"
boto3 = "~1.9.168"
django-registration = "~3.0.1"
django-ical = "=1.5"
django-notifications-hq = "~1.5.0"
elasticsearch = "<6"
django-markdown-filter = "~0.0.1"
django-node-assets = "~0.9.4"
factory-boy = "~2.12.0"
django-allow-cidr = "~0.3.1"
django-postman = "~3.6.2"
django-ajax-selects = "~1.7.1"
django-markdownify = "~0.8.0"
psycopg2-binary = "~2.8.2"
django = "~2.2.2"
numpy = "~1.16.4"
geopy = "~1.20.0"
django-haystack = { git = "https://github.com/andrlik/django-haystack.git", rev = "de8279680f74380085f61e247b2c4964c39db8e8" }
django-scheduler = { git = "https://github.com/andrlik/django-scheduler.git", rev = "4ae1fcf09065f0c9b7380685772549c70d2e78ff" }
sentry-sdk = "^0.9.0"
python-gitlab = "^1.11"
django-keybase-proofs = "^0.0.8"
django-allauth-2fa = "^0.7.0"
drf-extensions = "^0.5.0"
drf-writable-nested = "^0.5.1"
drf_yasg = "^1.17"
django-oauth-toolkit = "^1.2"
semantic_version = "^2.8"
rst_to_md = {git = "https://github.com/sixty-north/rst_to_md"}
ipython = "^7.9"

[tool.poetry.dev-dependencies]
sphinx = "~2.1.1"
pytest = "~4.6.3"
pytest-sugar = "~0.9.2"
pytest-cov = "~2.7.1"
pytest-xdist = "~1.28.0"
flake8 = "^3.7.7"
coverage = "~4.5.3"
django-test-plus = "~1.2.0"
django-coverage-plugin = "~1.6.0"
pytest-env = "~0.6.2"
bumpversion = "~0.5.3"
pypom = "~2.2.0"
pytest-selenium = "~1.16.0"
axe-selenium-python = "~2.1.6"
isort = "^4.3"
pytest-django = "^3.5"
django-debug-toolbar = "^1.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
